Add simpler segmentation algorithm
Implemented a simpler segmentation algorithm in a new function: segmentation_simpler
In this one, the initial mesh geometry is also passed to vtkPolyDataEdgeConnectivityFilter. This allows for simply passing the RegionId info when probing the input data, instead of looking it up later with cellLocator, etc. The runtime is also reduced. There are two disadvantages though (there is no free lunch!):
1. The boundaries of the delineated regions can have some irregularities. However, those irregularities are of the same order as the ones that exist anyway where different separatrices meet (i.e., near critical points).
2. It could identify individual mesh cells as regions if the cells are smaller than “dist”. This might be a serious issue for irregular meshes.